requirements.txt
pydantic
langchain
python-dotenv
langchain_groq
langchain_core
langchain_community
arxiv
wikipedia
Ipython
unicorn
streamlit
fastapi
